multi-account: made `algorand_public_key()` return `size_t` type instead of `int`.

+++ src/algo_keys.h
@@ -1,4 +1,4 @@
 #include "cx.h"
 
 void algorand_key_derive(uint32_t accountId, cx_ecfp_private_key_t *privateKey);
-int algorand_public_key(const cx_ecfp_private_key_t *privateKey, uint8_t *buf);
+size_t algorand_public_key(const cx_ecfp_private_key_t *privateKey, uint8_t *buf);
